consistently return "memory_usage" as double.

+++ src/KOKKOS/atom_vec_atomic_kokkos.cpp
@@ -879,9 +879,9 @@ void AtomVecAtomicKokkos::write_data(FILE *fp, int n, double **buf)
    return # of bytes of allocated memory
 ------------------------------------------------------------------------- */
 
-bigint AtomVecAtomicKokkos::memory_usage()
+double AtomVecAtomicKokkos::memory_usage()
 {
-  bigint bytes = 0;
+  double bytes = 0;
 
   if (atom->memcheck("tag")) bytes += memory->usage(tag,nmax);
   if (atom->memcheck("type")) bytes += memory->usage(type,nmax);
