+++ src/KOKKOS/atom_vec_full_kokkos.cpp
@@ -1588,9 +1588,9 @@ int AtomVecFullKokkos::write_data_hybrid(FILE *fp, double *buf)
    return # of bytes of allocated memory
 ------------------------------------------------------------------------- */
 
-bigint AtomVecFullKokkos::memory_usage()
+double AtomVecFullKokkos::memory_usage()
 {
-  bigint bytes = 0;
+  double bytes = 0;
 
   if (atom->memcheck("tag")) bytes += memory->usage(tag,nmax);
   if (atom->memcheck("type")) bytes += memory->usage(type,nmax);
